Added CompAnimatedSprite
